refactor(config): split config.go into per-domain files
Extract the sub-domain types and helpers that had accumulated in
config/config.go (~750 lines) into one-file-per-domain siblings,
mirroring the existing cache.go / budget.go / fallback.go convention.

Each new file holds its domain's types and any helpers co-located:
  server.go     — ServerConfig, NormalizeBasePath, JoinBasePath,
                  body-size constants and parsing
  models.go     — ModelsConfig, ConfiguredProviderModelsMode helpers
  storage.go    — StorageConfig family, BackendConfig
  guardrails.go — GuardrailsConfig, rules, prompt/altering settings
  resilience.go — Retry/CircuitBreaker/ResilienceConfig + Raw* overlays
  providers.go  — RawProviderConfig
  logging.go    — LogConfig
  usage.go      — UsageConfig
  admin.go      — AdminConfig
  http.go       — HTTPConfig
  workflows.go  — WorkflowsConfig
  metrics.go    — MetricsConfig

config.go now holds only the top-level Config struct, LoadResult,
buildDefaultConfig, Load, and applyYAML.

Move-only refactor: no symbol renames, no behavior changes. All callers
across the 15+ packages that import gomodel/config keep their existing
imports unchanged.
