+++ config/http.go
@@ -1,0 +1,12 @@
+package config
+
+// HTTPConfig holds HTTP client configuration for upstream API requests.
+// These values are also readable via the HTTP_TIMEOUT and HTTP_RESPONSE_HEADER_TIMEOUT
+// environment variables in internal/httpclient/client.go.
+type HTTPConfig struct {
+	// Timeout is the overall HTTP request timeout in seconds (default: 600)
+	Timeout int `yaml:"timeout" env:"HTTP_TIMEOUT"`
+
+	// ResponseHeaderTimeout is the time to wait for response headers in seconds (default: 600)
+	ResponseHeaderTimeout int `yaml:"response_header_timeout" env:"HTTP_RESPONSE_HEADER_TIMEOUT"`
+}
