+++ config/metrics.go
@@ -1,0 +1,12 @@
+package config
+
+// MetricsConfig holds observability configuration for Prometheus metrics
+type MetricsConfig struct {
+	// Enabled controls whether Prometheus metrics are collected and exposed
+	// Default: false
+	Enabled bool `yaml:"enabled" env:"METRICS_ENABLED"`
+
+	// Endpoint is the HTTP path where metrics are exposed
+	// Default: "/metrics"
+	Endpoint string `yaml:"endpoint" env:"METRICS_ENDPOINT"`
+}
